+++ src/observers/TelemetryObserver.ts
@@ -5,7 +5,7 @@
 
 import { PlatformInformation } from "../platform";
 import { BaseEvent, PackageInstallation, InstallationFailure, InstallationSuccess, OmnisharpDelayTrackerEventMeasures, OmnisharpStart, TestExecutionCountReport, TelemetryEventWithMeasures } from "../omnisharp/loggingEvents";
-import { PackageError } from "../packageManager/packages";
+import { PackageError } from "../packageManager/Package";
 
 export interface ITelemetryReporter {
     sendTelemetryEvent(eventName: string, properties?: { [key: string]: string }, measures?: { [key: string]: number }): void;
