+++ src/packageManager/PackageManager.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { PlatformInformation } from "../platform";
-import { Package, PackageError, NestedError } from './packages';
+import { Package, PackageError, NestedError } from './Package';
 import { DownloadFile } from './FileDownloader';
 import { InstallPackage } from './ZipInstaller';
 import { EventStream } from '../EventStream';
