+++ src/tools/UpdatePackageDependencies.ts
@@ -5,7 +5,7 @@
 
 import * as fs from 'fs';
 import * as os from 'os';
-import { Package } from '../packageManager/packages';
+import { Package } from '../packageManager/Package';
 
 
 interface PackageJSONFile
