+++ test/unitTests/logging/TelemetryObserver.test.ts
@@ -7,7 +7,7 @@ import { TelemetryObserver } from '../../../src/observers/TelemetryObserver';
 import { PlatformInformation } from '../../../src/platform';
 import { PackageInstallation, InstallationFailure, InstallationSuccess, TestExecutionCountReport, TelemetryEventWithMeasures, OmnisharpDelayTrackerEventMeasures, OmnisharpStart } from '../../../src/omnisharp/loggingEvents';
 import { getNullTelemetryReporter } from '../testAssets/Fakes';
-import { PackageError, Package } from '../../../src/packageManager/packages';
+import { PackageError, Package } from '../../../src/packageManager/Package';
 
 const chai = require('chai');
 chai.use(require('chai-arrays'));
